requirements.txt
python-telegram-bot==13.15
pandas
ta
openai
python-dotenv
requests
